Don't forward events to the cache stream
RE:
https://github.com/GoogleCloudPlatform/gcloud-node/issues/996#issuecomment-163139386

Previously, we were forwarding events from a request stream to
a cache stream. When the request is confirmed "good to go", the
cache stream is piped to the user's stream. The thought was the
events would be forwarded, too. It kind of worked, but had some
edge cases.

`streamForward` has always been tricky. It turns out it's not
all that helpful, but more confusing in practice. It's now
removed from use in this library. Instead, we are now forwarding
events manually.